@@ -1,14 +1,15 @@
 name: extended_image
 description: Official extension image, support placeholder(loading)/ failed state, cache network, zoom/pan, photo view, slide out page, editor(crop,rotate,flip), painting etc.
-version: 8.0.2
+version: 8.1.1
 homepage: https://github.com/fluttercandies/extended_image
 
 environment:
   sdk: '>=2.18.0 <4.0.0'
-  flutter: '>=3.10.0'
+  flutter: '>=3.13.0'
 
 dependencies:
-  extended_image_library: ^3.4.0
+  extended_image_library: ^3.6.0
   flutter:
     sdk: flutter
   meta: ^1.7.0
+dependency_overrides:
